+++ gemini-android/src/main/java/com/haroncode/gemini/android/binder/AndroidConnectionBinder.kt
@@ -1,25 +1,25 @@
 package com.haroncode.gemini.android.binder
 
 import android.os.Looper
-import com.haroncode.gemini.core.Binder
+import com.haroncode.gemini.core.ConnectionBinder
 import com.haroncode.gemini.core.ConnectionRule
 
-class AndroidLifecycleBinder(
-    private val binder: Binder
-) : Binder {
+class AndroidConnectionBinder(
+    private val connectionBinder: ConnectionBinder
+) : ConnectionBinder {
 
     override fun bind(connectionRule: ConnectionRule) {
         check(Looper.myLooper() == Looper.getMainLooper()) { "Method bind() can invoked only in main thread" }
-        binder.bind(connectionRule)
+        connectionBinder.bind(connectionRule)
     }
 
     override fun dispose() {
         check(Looper.myLooper() == Looper.getMainLooper()) { "Method dispose() can invoked only in main thread" }
-        binder.dispose()
+        connectionBinder.dispose()
     }
 
     override fun isDisposed(): Boolean {
         check(Looper.myLooper() == Looper.getMainLooper()) { "Method isDisposed() can invoked only in main thread" }
-        return binder.isDisposed
+        return connectionBinder.isDisposed
     }
 }
