+++ gemini-android/src/main/java/com/haroncode/gemini/android/binder/BindingStrategy.kt
@@ -1,9 +1,9 @@
 package com.haroncode.gemini.android.binder
 
 import androidx.lifecycle.LifecycleObserver
-import com.haroncode.gemini.binder.BinderLifecycle
+import com.haroncode.gemini.core.StoreLifecycle
 
-typealias StateSender = (BinderLifecycle.State) -> Unit
+typealias StateSender = (StoreLifecycle.State) -> Unit
 
 interface BindingStrategy {
