+++ gemini-android/src/main/java/com/haroncode/gemini/android/binder/ViewBinding.kt
@@ -1,8 +1,8 @@
 package com.haroncode.gemini.android.binder
 
-import com.haroncode.gemini.core.Binder
+import com.haroncode.gemini.core.ConnectionBinder
 
 interface ViewBinding<View : Any> {
 
-    fun onCreate(view: View, binder: Binder)
+    fun onCreate(view: View, binder: ConnectionBinder)
 }
